added link to all news, created checks for controllers in admin, modified listing output on news and comments to show customer/admin name instead of id

+++ Inchoo/NewsWithComments/Controller/Adminhtml/News/Delete.php
@@ -21,9 +21,12 @@ class Delete extends Action
 
     public function execute()
     {
-        $ids = $this->_request->getParam('selected');
-        $message = $this->newsRepository->deleteNews($ids);
-        $this->messageManager->addNoticeMessage($message);
+        if ($ids = $this->_request->getParam('selected')) {
+            $message = $this->newsRepository->deleteNews($ids);
+            $this->messageManager->addNoticeMessage($message);
+            return $this->_redirect('news/news');
+        }
+
         return $this->_redirect('news/news');
     }
 }
